feat(ipfs): move around lib files, put snapshot and nance.ts lib stuff into folders, use @ imports when possible
