+++ ui/components/nance/ProposalInfo.tsx
@@ -15,7 +15,7 @@ import toastStyle from '../../lib/marketplace/marketplace-utils/toastConfig'
 import { NANCE_SPACE_NAME } from '../../lib/nance/constants'
 import useAccount from '../../lib/nance/useAccountAddress'
 import { useSignProposal } from '../../lib/nance/useSignProposal'
-import { SnapshotGraphqlProposalVotingInfo } from '../../lib/snapshot'
+import { SnapshotGraphqlProposalVotingInfo } from '@/lib/snapshot'
 import { AddressLink } from './AddressLink'
 import ProposalStatusIcon from './ProposalStatusIcon'
 import RequestingTokensOfProposal from './RequestingTokensOfProposal'
