+++ ui/components/nance/RequestBudgetActionForm.tsx
@@ -5,7 +5,7 @@ import AddressForm from './form/AddressForm'
 import DiscordUserIdForm from './form/DiscordUserIdForm'
 import SafeTokenForm from './form/SafeTokenForm'
 import StringForm from './form/StringForm'
-import { useEffect } from "react"
+// import { useEffect } from "react"
 
 export default function RequestBudgetActionForm() {
   // form
